pre-last commit

+++ console/migrations/m210122_153400_create_apples_table.php
@@ -18,7 +18,7 @@ class m210122_153400_create_apples_table extends Migration
             'createDate' => $this->dateTime(),
             'fallToGroundDate' => $this->dateTime(),
             'appleStatus' => $this->integer(),
-            'eatingProcent' => $this->integer(),
+            'eatingProcent' => $this->float(),
         ]);
     }
